test(merge-lease): give the lease spec its own file
The lease's whole specification -- 11 cases -- lived inside
gate-dispatch.test.mjs, a file about the gate dispatcher, which is not the
lease and does not call it. Move them unchanged into scripts/merge-lease.test.mjs
and run it from the gate plan's existing node --test member.

The cases are byte-identical; only their fixture environment is local to the
new file, where it neutralises MERGE_LEASE_* rather than the gate's namespace.
