+++ public-snapshot.json
@@ -115,7 +115,7 @@
     { "glob": "scripts/gate-worker/bench-dbtest-concurrency.sh", "purpose": "gate worker: dbtest concurrency benchmark over one fixed commit" },
     { "glob": "scripts/gate-worker/bench-postgres.sh", "purpose": "gate worker: throwaway-postgres settings benchmark over one fixed commit" },
     { "glob": "scripts/gate-worker/gate-dispatch.sh", "purpose": "gate worker: slot-rationed dispatch across the local machine and the worker" },
-    { "glob": "scripts/gate-worker/gate-dispatch.test.mjs", "purpose": "gate worker and merge lease: slot-lock concurrency, dispatcher retry and exit-code behavior, and origin-ref lease regression tests" },
+    { "glob": "scripts/gate-worker/gate-dispatch.test.mjs", "purpose": "gate worker: slot-lock concurrency, dispatcher retry and exit-code behavior" },
     { "glob": "scripts/gate-worker/gate-worker.test.mjs", "purpose": "gate worker: credential red-line parity, remote-value allowlists, exact-ref transport retries and run-gate no-verdict regression tests" },
     { "glob": "scripts/gate-worker/lib.sh", "purpose": "gate worker: shared repository-name derivation for the local-side scripts" },
     { "glob": "scripts/gate-worker/mirror-push.sh", "purpose": "gate worker: mirror push, the only path code reaches the worker by" },
@@ -132,6 +132,7 @@
     { "glob": "scripts/merge-gate-profile.test.mjs", "purpose": "merge-gate profile classifier fixtures" },
     { "glob": "scripts/merge-gate-parallel.test.mjs", "purpose": "merge-gate concurrent-group failure reporting fixtures" },
     { "glob": "scripts/merge-lease.sh", "purpose": "global exact-head merge delivery lease backed by an atomic origin ref" },
+    { "glob": "scripts/merge-lease.test.mjs", "purpose": "merge lease: mutual exclusion, release outcomes, status and the machine and human steal boundaries against a real origin ref" },
     { "glob": "scripts/merge-integrator-real-checks.mjs", "purpose": "redacted real-infrastructure merge-integrator checks" },
     { "glob": "scripts/merge-integrator-system-test.mjs", "purpose": "redacted end-to-end merge-integrator system harness" },
     { "glob": "scripts/os-isolation/patch-runner-plists.sh", "purpose": "OS isolation: launchd RUNNER_RUN_AS_PREFIX wiring" },
